% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{getkmer}
\alias{getkmer}
\title{Compare the kmer difference between 2 sets of sequences}
\usage{
getkmer(
  targetfile = NULL,
  genomename = "mm10",
  k = 6,
  genes1,
  genes2,
  feature,
  radius = 1000
)
}
\arguments{
\item{targetfile}{The directory of the file indicating the gene regions
whose kmer need to be compared between the 2 sequence sets, not necessary
to be extactly the gene body region between TSS and TTS sites. Columns
named as chr, start, end, strand, and gene_id are required. If it is NULL,
all the gene regions defined together by the parameters \code{genomename},
\code{feature} and \code{radius} will be analyzed.}

\item{genomename}{Specify the genome of the genes to be analyzed, when the
parameter \code{targetfile} is NULL.}

\item{k}{The length of the kmer to be analyzed. Default is 6, meaning 6-mer
will be analyzed.}

\item{genes1}{The symbols of genes in set 1 whose kmers need to be compared
with that of set 2. The regions indicated by the parameter
\code{targetfile} or \code{genomename} etc will be used for these genes
only if they belong to the genes indicated by this parameter
\code{genes1}.}

\item{genes2}{The symbols of genes in set 2 whose kmers need to be compared
with that of set 1. Similar to the parameter \code{genes1}.}

\item{feature}{If the parameter \code{targetfile} is NULL, while the
parameter \code{genomename} is defined. This parameter \code{feature} can
be used to further select regions from the genes indicated by
\code{genomename}. Can choose from 'promoter', 'end', and 'genebody'. If
it is 'promoter' or 'end', another parameter \code{radius} is needed to
define the radius of the promoter or end region centering around the TSS
or TTS site.}

\item{radius}{A numberic value needed to define the radius length of the
gene promoter or end region if the parameter \code{feature} is set as
'promoter' or 'end'.}
}
\value{
A data.frame indicating the kmer frquency ratios between sequence
set 1 and set 2, as well as p-values (calculated with Fisher's test) and
adjusted p-values (adjusted using the Benjamini & Hochberg method).
}
\description{
Compare the kmer difference between 2 sets of sequences, reporting the ratio
between the kmer frequencies of the 2 sequence sets, the p-value and
adjusted p-value of the difference.
}
