% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{intercompare}
\alias{intercompare}
\title{Compare gene elongation rates between 2 different conditions}
\usage{
intercompare(inferresmat, groupnames)
}
\arguments{
\item{inferresmat}{A data.frame recording the elongation rates for the genes
in different conditions. Can be generated using the function
\code{orgintercompareinput}.}

\item{groupnames}{A vector with the elements corresponding to the condition
names of the rate columns in the data.frame indicated by the parameter
\code{inferresmat}.}
}
\value{
A data.frame with one row for one gene and the columns are the
average transcription elongation rates for each condition (log2
transformed), the log2FC, the p-value and adjusted p-value (adjusted by
Benjamini & Hochberg method) of the difference.
}
\description{
Compare gene elongation rate difference for each gene between 2 different
conditions. For each condition, if the number of replicates is < 3, Fisher's
test will be used to calculate the p-values and adjusted p-values of the
rate differences, while if the number of replicates is >= 3, Wilcox test and
limma will be used to calculate 2 sets of p-values and adjusted p-values
(Wilcox test based and limma based).
}
