% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{intracompare}
\alias{intracompare}
\title{Divide genes into different quantile groups according to their rates}
\usage{
intracompare(inferres, targetgenes = NULL, quantilenum = 4)
}
\arguments{
\item{inferres}{The result list generated by the \code{calrate} function or
\code{mcalrate} function.}

\item{targetgenes}{Which genes need to be divided into quantile groups. If
it is NULL (The default value), all genes provided by the parameter
\code{inferres} will be analyzed. If its value is 'significant', only
genes judged as significant genes by \code{calrate} or \code{mcalrate}
will be analyzed. While the the value of this parameter can also be a
vector with gene symbols as elements and only genes covered by this
vector will be analyzed.}

\item{quantilenum}{How many quantile groups need to be divided. Default value
is 4.}
}
\value{
A list with 2 data.frames as elements. One of them records the
absolute elongation rates of the genes (bp/min) and divide the genes into
several quantile groups with genes in quantile1 group having the lowest
rates. The other data.frame records the relative elongation rates of the
genes (percent of gene body/min) and divide them into quantile groups. If
the value of parameter \code{inferres} is from \code{mcalrate}, the
absolute and relative rates of the genes are the mean of all the Pro-seq
or Gro-seq pairs.
}
\description{
Divide genes into different quantile groups according to their rates
inferred from the \code{calrate} function, or \code{mcalrate} function.
}
