% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{makegenelist}
\alias{makegenelist}
\title{Generate a GRanges object of UCSC known genes in sepcific genome}
\usage{
makegenelist(genomename = "mm10", save = TRUE)
}
\arguments{
\item{genomename}{The name of the genome, can be 'mm10' for mouse or 'hg38'
for human.}

\item{save}{Whether save the result GRanges as a file in work directory,
default is TRUE.}
}
\value{
A GRanges object of UCSC known genes, with several gene information
provided, including seqnames, ranges, strands, gene symbols, distance to
the nearest upstream gene, distance to the nearest downstream gene, gene
GC content, and gene exon coverage. Genes overlapping with each other will
not be included.
}
\description{
Generate a GRanges object of UCSC known genes in specific genome.
}
