% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{orgintercompareinput}
\alias{orgintercompareinput}
\title{Organize the results from the function \code{mcalrate} to a data.frame}
\usage{
orgintercompareinput(reslist, targetgenes = NULL)
}
\arguments{
\item{reslist}{The result list generated by the function \code{mcalrate}.}

\item{targetgenes}{Which genes need to be included in the output of this
function. If it is NULL (The default value), all genes provided by the
parameter \code{reslist} will be analyzed. If its value is 'significant',
only genes judged as significant genes by \code{mcalrate} will be
included. While the the value of this parameter can also be a vector with
gene symbols as elements and only genes coverred by this vector will be
included.}
}
\value{
A data.frame with the 1st column as gene symbols and the other
columns as gene elongation rates inferred for each Pro-seq or Gro-seq
pair. This data.frame can be used as input for the function
\code{intercompare}.
}
\description{
Organize the results calculated by the function \code{mcalrate} to a
data.frame with the 1st column as gene symbols and the other columns as gene
elongation rates inferred for each Pro-seq or Gro-seq pair. This data.frame
can be used as input for the function \code{intercompare}.
}
