% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supps.R
\name{plotprocessing}
\alias{plotprocessing}
\title{Process the results of the functions \code{metaplot} and \code{mmetaplot}}
\usage{
plotprocessing(
  fwdlist,
  revlist,
  groupnames,
  lineposes,
  labels,
  title,
  cutoff = 0.01,
  titlesize = 17,
  textsize = 16
)
}
\arguments{
\item{fwdlist}{A list containing the metagene FPM values from the function
\code{metaplot} or \code{mmetaplot}'s result list. Each element in it is
an FPM value vector contained in the original result list, and its slot
name there is "TSSfwdFPMmeans", "TTSfwdFPMmeans", "GeneBodyfwdFPMmeans",
or "combinefwdFPMmeans". It is the FPM value vector for the metagene's
sense strand and should match the corresponding list element transferred
to another parameter \code{revlist}, which is the FPM value vector for the
metagene's antisense strand. For the different elements of \code{fwdlist},
they should come from the same regions, such as the TSS region, the TTS
region, the gene body region, etc. Their difference is that they may from
different experimental conditions, such as the TSS region's metagene data
of wildtype samples and gene knockout samples, respectively.}

\item{revlist}{A list containing the metagene FPM values from the function
\code{metaplot} or \code{mmetaplot}'s result list. Each element in it is
an FPM value vector contained in the original result list, and its slot
name there is "TSSrevFPMmeans", "TTSrevFPMmeans", "GeneBodyrevFPMmeans",
or "combinerevFPMmeans". It is the FPM value vector for the metagene's
antisense strand and should match the list element transferred to another
parameter \code{fwdlist}, which is the FPM value vector for the metagene's
sense strand. For the different elements of \code{fwdlist}, they should
come from the same regions, such as the TSS region, the TTS region, the
gene body region, etc. Their difference is that they may from different
experimental conditions, such as the TSS region's metagene antisense FPM
values of wildtype samples and gene knockout samples, respectively.}

\item{groupnames}{A vector with elements as strings to show the different
elements' experimental conditions in \code{fwdlist} and \code{revlis}.}

\item{lineposes}{A vector with elements as integers to show the coordinates
of the TSS and/or TTS points in the metagene. These coordinates use the
beginning of the metagene x-coordinate as 1.}

\item{labels}{A vector with elements as strings to show the x-axis labels
for the beginning and end of the metagene's x-axis and for the positions
indicated by the parameter \code{lineposes}. The labels should be ordered
following their coordinates along the x-axis.}

\item{title}{The title for the metagene plot.}

\item{cutoff}{To remove the extremely large FPM outliers from the original
metagene plot, this parameter needs to be set. The default value is 0.01,
which means the 1 - 1\% (0.01) = 99\% quantile of the metagene FPMs will be
defined as their maximum, and any larger ones will be reduced.}

\item{titlesize}{The font size for the plot title. Default is 17.}

\item{textsize}{The font size for the plot texts. Default is 16.}
}
\value{
The processed metagene plot with the FPM outliers removed.
}
\description{
Process the metagene results from \code{metaplot} and \code{mmetaplot} and
remove the extremely large FPM outliers in the original metagene plots.
}
