% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mcalrate}
\alias{mcalrate}
\title{Calculate gene elongation rate for multiple pairs of Pro-seq or Gro-seq data}
\usage{
mcalrate(
  time1files,
  time2files,
  targetfile,
  genomename = NULL,
  times,
  strandmethod = 1,
  threads = 1,
  mergerefs = TRUE,
  mergecases = FALSE,
  savegenenames,
  plotgenenames = TRUE,
  genelencutoff = 70000,
  fpkmcutoff = 1
)
}
\arguments{
\item{time1files}{The reference Pro-seq/Gro-seq bam file names,
corresponding to 0min. Should be a vector with elements as strings
indicating the directores of the files.}

\item{time2files}{The treatment Pro-seq/Gro-seq bam file names,
corresponding to several specific time points (e.g. 15min if DRB was used
for 15min). Should be a vector with elements as strings indicating the
directores of the files.}

\item{targetfile}{The genes whose transcriptional elongation rate need to be
analyzed. If it is NULL, all the genes longer than at least 4000bp in the
genome specified by the parameter \code{genomename} will be analyzed. If
provied by the user, columns named as chr, start, end, strand, and gene_id
are required. All genes should have a length greater than at least 4000bp.}

\item{genomename}{Specify the genome of the genes to be analyzed, when the
parameter \code{targetfile} is NULL.}

\item{strandmethod}{The strand specific method used when preparing the
sequencing library, can be 1 for directional ligation method and 2 for
dUTP method. If the sample is sequenced using a single strand method, set
it as 0.}

\item{threads}{Number of threads to do the parallelization. Default is 1.}

\item{mergerefs}{Whether merge the data of all time1 reference files
together to one reference, and then use it as a uniform reference for all
\code{time2files}. Default is TRUE.}

\item{mergecases}{Whether merge the data of all time2 case files together to
one case file. Default is FALSE.}

\item{savegenenames}{For which genes their concrete rate inference results
need to be saved, including the binned ratio values between their data in
\code{time2files} and \code{time1files}, the extended ratio values for
specific bins, etc.}

\item{plotgenenames}{Whether to plot the binned and extended ratio values
for the genes provided by \code{savegenenames}}

\item{genelencutoff}{The cutoff on gene length (bp). Only genes longer than
this cutoff will be considerred. If its value is NULL or less than 4000,
genes longer than 4000bp will be considerred. Default is 70000.}

\item{fpkmcutoff}{The cutoff on gene FPKM value. Only genes with an FPKM
value greater than the cutoff in time1file will be considerred. Default is
1.}

\item{time}{The time differences between time1files and time2files, using
min as its unit. Should be a vector with each element as the time
difference between the time1file and time2file presented in the
corresponding positions in \code{time1files} vector and \code{time2files}
vector.}
}
\value{
A list with each of its element as a sub-list recording the
transcription elongation rate inferred for each pair of Pro-seq or Gro-seq
data using the LSS method, as well as other information, such as the
significance of result.
}
\description{
Calculate gene elongation rate for multiple pairs of Pro-seq data, or
Gro-seq data, using an LSS (least sum of square) based method.
}
