% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{metaplot}
\alias{metaplot}
\title{Generate metagene plots from a bam file}
\usage{
metaplot(
  metafile,
  targetgenefile,
  genomename = NULL,
  tssradius = c(2000, 1000, 500),
  ttsradius = c(2000, 1000, 500),
  genebodylen = 4000,
  label = NULL,
  strandmethod = 1,
  threads = 1,
  savegenenames,
  plotgenenames = TRUE,
  plotfig = TRUE,
  genelencutoff = NULL,
  fpkmcutoff = 1
)
}
\arguments{
\item{metafile}{The bam file need to generate metagene plots. Can be a
string indicating the directory of the file, or a GAlignmentPairs object,
or a GRanges object from the original bam file.}

\item{targetgenefile}{The genes whose FPM values need to be merged together
to generate a metagene plot. If it is NULL, all the genes in the genome
specified by the parameter \code{genomename} will be analyzed. If provied
by the user, columns named as chr, start, end, strand, and gene_id are
required. All genes should be longer than the maximum value of the
parameters \code{tssradius}, \code{ttsradius}, and \code{genelencutoff}.}

\item{genomename}{Specify the genome of the genes to be analyzed, when the
parameter \code{targetgenefile} is NULL.}

\item{tssradius}{If want to plot the metagene in TSS region, set this
parameter as a vector with each element corresponding to a radius of the
TSS region centering around the TSS site. Then, for each radius value, a
metagene plot will be generated.}

\item{ttsradius}{If want to plot the metagene in TTS region, set this
parameter as a vector with each element corresponding to a radius of the
TTS region centering around the TTS site. Then, for each radius value, a
metagene plot will be generated.}

\item{genebodylen}{If want to plot the metagene in gene body region, set
this parameter as a specific numeric value (bp). Because different genes
have different gene body length, before merge their FPM values to a
metagene, their gene body length will be scaled to a uniform length, which
is set by this parameter \code{genebodylen}. Genes with a longer length
will be compressed to this gene length, while the ones with a shorter
length will be exteneded.}

\item{label}{A string that will be included in the title of the metagene
plot to indicate the experimental condition.}

\item{strandmethod}{The strand specific method used when preparing the
sequencing library, can be 1 for directional ligation method and 2 for
dUTP method. If the sample is sequenced using a single strand method, set
it as 0.}

\item{threads}{Number of threads to do the parallelization. Default is 1.}

\item{savegenenames}{For which genes their concrete FPM value for each bp
position need to be saved, or plotted.}

\item{plotgenenames}{Whether to plot the FPM value for each bp position for
the genes provided by the parameter \code{savegenenames}.}

\item{plotfig}{Whether plot the metagenes or not. Default value is TRUE. For
the genes indicated by \code{savegenenames}, only if both the paramters
\code{plotgenenames} and \code{plotfig} are TRUE, they will be plotted.}

\item{genelencutoff}{The cutoff on gene length (bp). The default value is
NULL, but if it is set, only genes longer than this cutoff will be
considerred.}

\item{fpkmcutoff}{The cutoff on gene FPKM value. Only genes with an FPKM
value greater than the cutoff will be considerred. Default is 1.}
}
\value{
Will generate several metagene plots as well as a list containing
the information of the FPKM values on specific regions for each gene, the
concrete FPM value on each bp position for the metagene, and the genes
indicated by the parameter \code{savegenenames}, etc.
}
\description{
Generate metagene plots from a bam file, as well as other information
including the FPM value of each bp position in the metagene plot, FPKM value
in specific TSS, TTS, or gene body regions of each gene, etc.
}
